requirements.txt
streamlit
pandas
supabase
python-dotenv